writer template
