+++ mabel/operators/google/__init__.py
@@ -1,2 +1,2 @@
-from .save_snapshot_to_google_storage_operator import SaveSnapshotToGoogleStorageOperator
+from .google_storage_batch_writer_operator import GoogleCloudStorageWriter
 from .stream_to_google_storage_operator import StreamToGoogleStorageOperator
