--- mabel/operators/google/save_snapshot_to_google_storage_operator.py
@@ -1,48 +1,0 @@
-import datetime
-from ..internals.base_operator import BaseOperator
-from ...data import BatchWriter
-from ...adapters.google import GoogleCloudStorageWriter
-from ...data.validator import Schema  # type:ignore
-
-
-class SaveSnapshotToGoogleStorageOperator(BaseOperator):
-
-    def __init__(
-            self,
-            *,
-            project: str,
-            dataset: str,
-            schema: Schema = None,
-            format: str = 'zstd',
-            date: datetime.date = None,
-            **kwargs):
-        """
-        This is a wrapper around orwell.data.writers.GoogleCloudStorage.
-
-        Parameters:
-            project: string
-            dataset: string
-            schema: Schema (optional)
-            format: string (optional)
-            date: date (optional)
-        """
-        super().__init__()
-        self.writer = BatchWriter(
-                inner_writer=GoogleCloudStorageWriter,
-                project=project,
-                dataset=dataset,
-                schema=schema,
-                format=format,
-                date=date,
-                **kwargs)
-
-    def execute(self, data: dict = {}, context: dict = {}):
-        self.writer.append(data)
-        return data, context
-
-    def finalize(self, context):
-        self.writer.finalize()
-        return context
-
-    def __del__(self):
-        self.writer.finalize()
