+++ mabel/operators/internals/base_writer_operator.py
@@ -1,0 +1,29 @@
+from ..internals.base_operator import BaseOperator
+
+
+class BaseWriterOperator(BaseOperator):
+
+    def __init__(
+            self,
+            *,
+            writer,
+            inner_writer,
+            **kwargs):
+        """
+        Writers should extend this class
+        """
+        super().__init__(**kwargs)
+        self.writer = writer(
+                inner_writer=inner_writer,
+                **kwargs)
+
+    def execute(self, data: dict = {}, context: dict = {}):
+        self.writer.append(data)
+        return data, context
+
+    def finalize(self, context):
+        self.writer.finalize()
+        return context
+
+    def __del__(self):
+        self.writer.finalize()
