--- mabel/operators/local/__init__.py
@@ -1,2 +1,0 @@
-from .save_snapshot_to_disk_operator import SaveSnapshotToDiskOperator
-from .stream_to_disk_operator import StreamToDiskOperator
