+++ mabel/operators/minio/minio_batch_writer_operator.py
@@ -1,10 +1,9 @@
-from ..internals.base_operator import BaseOperator
+from ..internals.base_writer_operator import BaseWriterOperator
 from ...data import BatchWriter
 from ...adapters.minio import MinIoWriter
-from ...data.validator import Schema  # type:ignore
 
 
-class SaveSnapshotToMinIoOperator(BaseOperator):
+class MinIoBatchWriterOperator(BaseWriterOperator):
 
     def __init__(
             self,
@@ -16,22 +15,12 @@ class SaveSnapshotToMinIoOperator(BaseOperator):
             **kwargs):
         """
         """
-        super().__init__()
         kwargs['end_point'] = end_point
         kwargs['access_key'] = access_key
         kwargs['secret_key'] = secret_key
         kwargs['secure'] = secure
-        self.writer = BatchWriter(
+
+        super().__init__(
+                writer=BatchWriter,
                 inner_writer=MinIoWriter,
                 **kwargs)
-
-    def execute(self, data: dict = {}, context: dict = {}):
-        self.writer.append(data)
-        return data, context
-
-    def finalize(self, context):
-        self.writer.finalize()
-        return context
-
-    def __del__(self):
-        self.writer.finalize()
