+++ tests/test_operator_save_snapshop_to_disk_operator.py
@@ -2,7 +2,7 @@ import os
 import sys
 import shutil
 sys.path.insert(1, os.path.join(sys.path[0], '..'))
-from mabel.operators.local import SaveSnapshotToDiskOperator
+from mabel.operators.disk import DiskBatchWriterOperator
 try:
     from rich import traceback
     traceback.install()
@@ -14,7 +14,7 @@ def test_save_to_disk_operator():
 
     shutil.rmtree('_temp', ignore_errors=True)
 
-    n = SaveSnapshotToDiskOperator(
+    n = DiskBatchWriterOperator(
             dataset="_temp/",
             format='zstd')
     n.execute(data={"this":"is", "a":"record"}, context={})
